refactor(v3): extract Server lifecycle + accessors from server.go (777→461)
Pure same-package relocation, no call-site or behavior changes:

- server.go (461): Server struct, NewServer, Dependencies, and all
  dependency-wiring setters
- server_lifecycle.go: StartMonitor, StartRecordingCacheEvicter,
  Shutdown, SetRuntimeContext
- server_accessors.go: config/status getters, auth middleware,
  data-path + OpenWebIF client helpers

The dependency-wiring setters are deliberately kept in server.go: they
touch raw Server state fields that TestModuleBoundaries_NoDirectServer-
StateFieldAccess confines to server.go/server_modules.go. Only the
lifecycle and accessor methods (zero forbidden-field access) move out.

23 funcs in, 23 funcs out (lossless). Verified by go build, go vet,
the full v3 suite incl. the module-boundary test, scoped golangci-lint,
goimports.
